+++ engine/src/resources/sounds.rs
@@ -7,13 +7,14 @@ use std::{
     time::Duration,
 };
 
+use anyhow::Result;
 use crossbeam::channel::{unbounded, Sender};
 use glam::{Quat, Vec3};
 use kira::{
-    manager::{backend::DefaultBackend, AudioManager, AudioManagerSettings},
+    manager::{backend::DefaultBackend, AudioManager, AudioManagerSettings, Capacities},
     sound::{
         static_sound::{StaticSoundData, StaticSoundHandle, StaticSoundSettings},
-        FromFileError, IntoOptionalRegion, PlaybackState,
+        FromFileError,
     },
     spatial::{
         emitter::{EmitterHandle, EmitterSettings},
@@ -26,11 +27,9 @@ use kira::{
 pub(crate) fn audio_server() -> Sender<AudioUpdate> {
     let (send, recv) = unbounded();
     thread::spawn(|| {
-        let mut audio_manager =
-            AudioManager::<DefaultBackend>::new(AudioManagerSettings::default()).unwrap();
-        let mut spacial_scene = audio_manager
-            .add_spatial_scene(SpatialSceneSettings::default())
-            .unwrap();
+        let (manager_settings, scene_settings) = SETTINGS.audio_settings().make();
+        let mut audio_manager = AudioManager::<DefaultBackend>::new(manager_settings).unwrap();
+        let mut spacial_scene = audio_manager.add_spatial_scene(scene_settings).unwrap();
         let recv = recv;
         loop {
             let update: AudioUpdate = recv.recv().unwrap();
@@ -56,15 +55,13 @@ pub(crate) fn audio_server() -> Sender<AudioUpdate> {
                         settings = settings.output_destination(&spatial_emitter);
                         let _ = emitter.set(spatial_emitter);
                     }
-                    let handle = audio_manager
-                        .play(StaticSoundData {
-                            sample_rate: sound.data.sample_rate,
-                            frames: sound.data.frames,
-                            settings,
-                        })
-                        .unwrap();
+                    let handle = audio_manager.play(StaticSoundData {
+                        sample_rate: sound.data.sample_rate,
+                        frames: sound.data.frames,
+                        settings,
+                    });
                     sound.handle.lock().take();
-                    let _ = sound.handle.lock().set(handle);
+                    let _ = sound.handle.lock().set(handle.map_err(|x| x.into()));
                 }
                 AudioUpdate::NewLayer(layer) => {
                     let _ = layer.listener.lock().set(
@@ -72,6 +69,11 @@ pub(crate) fn audio_server() -> Sender<AudioUpdate> {
                             .add_listener(Vec3::ZERO, Quat::IDENTITY, ListenerSettings::default())
                             .unwrap(),
                     );
+                }
+                AudioUpdate::SettingsChange(settings) => {
+                    let (manager_settings, scene_settings) = settings.make();
+                    audio_manager = AudioManager::<DefaultBackend>::new(manager_settings).unwrap();
+                    spacial_scene = audio_manager.add_spatial_scene(scene_settings).unwrap();
                 }
             };
         }
@@ -82,18 +84,138 @@ pub(crate) fn audio_server() -> Sender<AudioUpdate> {
 pub(crate) enum AudioUpdate {
     Play(Sound),
     NewLayer(Arc<Layer>),
+    SettingsChange(AudioSettings),
 }
 
-// future wrap
 pub use kira::{
-    dsp::*,
-    sound::{PlaybackRate, Region},
-    tween::{Easing, Tween},
-    StartTime, Volume,
+    dsp::Frame,
+    sound::{
+        EndPosition, IntoOptionalRegion, PlaybackPosition, PlaybackRate, PlaybackState, Region,
+    },
+    tween::Easing,
+    Volume,
 };
+
+/// "Inbetween"
+///
+/// Describes an interpolation between values.
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct Tween {
+    /// The duration of the interpolation.
+    pub duration: Duration,
+    /// The easing method used.
+    pub easing: Easing,
+}
+
+impl Default for Tween {
+    fn default() -> Self {
+        kira::tween::Tween::default().into()
+    }
+}
+
+impl From<kira::tween::Tween> for Tween {
+    fn from(value: kira::tween::Tween) -> Self {
+        Self {
+            duration: value.duration,
+            easing: value.easing,
+        }
+    }
+}
+impl From<Tween> for kira::tween::Tween {
+    fn from(value: Tween) -> Self {
+        Self {
+            duration: value.duration,
+            easing: value.easing,
+            start_time: kira::StartTime::Immediate,
+        }
+    }
+}
+
+#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Debug)]
+pub struct AudioSettings {
+    /// The limit of how many sounds can exist at the same time.
+    pub sound_capacity: usize,
+    /// The limit of how many sounds can be bound to objects to make them spatial.
+    pub object_bound_sound_capacity: usize,
+    /// The limit of how many scenes can play spatial sounds.
+    pub spatial_scene_capacity: usize,
+}
+
+impl AudioSettings {
+    /// Makes default audio settings.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Sets the maximum amount of sounds.
+    pub fn set_sound_capacities(&mut self, sound_capacity: usize) {
+        self.sound_capacity = sound_capacity;
+    }
+
+    /// Sets the maximum amount of sounds and returns self.
+    pub fn sound_capacity(mut self, sound_capacity: usize) -> Self {
+        self.sound_capacity = sound_capacity;
+        self
+    }
+
+    /// Sets the maximum amount of sounds you can bind on objects to make them spatial.
+    pub fn set_object_bound_sound_capacity(&mut self, sound_capacity: usize) {
+        self.object_bound_sound_capacity = sound_capacity;
+    }
+
+    /// Sets the maximum amount of sounds you can bind on objects to make them spatial and returns self.
+    pub fn object_bound_sound_capacity(mut self, sound_capacity: usize) -> Self {
+        self.object_bound_sound_capacity = sound_capacity;
+        self
+    }
+
+    /// Sets the maximum amount of scenes that can play spatial sounds.
+    pub fn set_spatial_scene_capacity(&mut self, scene_capacity: usize) {
+        self.spatial_scene_capacity = scene_capacity;
+    }
+
+    /// Sets the maximum amount of scenes that can play spatial sounds and returns self.
+    pub fn spatial_scene_capacity(mut self, scene_capacity: usize) -> Self {
+        self.spatial_scene_capacity = scene_capacity;
+        self
+    }
+
+    /// Converts these audio settings to the kira settings to be used when making or editing the settings.
+    pub(crate) fn make(&self) -> (AudioManagerSettings<DefaultBackend>, SpatialSceneSettings) {
+        let manager_settings = AudioManagerSettings {
+            capacities: Capacities {
+                command_capacity: 256,
+                sound_capacity: self.sound_capacity,
+                clock_capacity: 1,
+                spatial_scene_capacity: self.spatial_scene_capacity,
+                ..Default::default()
+            },
+            ..Default::default()
+        };
+        let scene_settings = SpatialSceneSettings::new()
+            .emitter_capacity(self.object_bound_sound_capacity)
+            .listener_capacity(self.spatial_scene_capacity);
+
+        (manager_settings, scene_settings)
+    }
+}
+
+impl Default for AudioSettings {
+    fn default() -> Self {
+        Self {
+            sound_capacity: 256,
+            object_bound_sound_capacity: 256,
+            spatial_scene_capacity: 8,
+        }
+    }
+}
+
 use parking_lot::Mutex;
 
-use crate::objects::{scenes::Layer, Object};
+use crate::{
+    objects::{scenes::Layer, Object},
+    SETTINGS,
+};
 
 use super::RESOURCES;
 
@@ -192,7 +314,7 @@ pub struct Sound {
     data: SoundData,
     settings: SoundSettings,
     emitter: Arc<Mutex<OnceLock<EmitterHandle>>>,
-    handle: Arc<Mutex<OnceLock<StaticSoundHandle>>>,
+    handle: Arc<Mutex<OnceLock<Result<StaticSoundHandle>>>>,
     object: Option<Object>,
 }
 
@@ -207,13 +329,23 @@ impl Sound {
             object: None,
         }
     }
+    /// Sets the settings of this sound.
+    pub fn set_settings(&mut self, settings: SoundSettings) {
+        self.settings = settings;
+    }
+
+    /// Returns the settings of this sounds.
+    pub fn settings(&self) -> SoundSettings {
+        self.settings
+    }
+
     /// Returns the data behind the sound.
     pub fn data(&self) -> &SoundData {
         &self.data
     }
     /// Returns the current playback state of the sound.
     pub fn state(&self) -> PlaybackState {
-        if let Some(handle) = self.handle.lock().get() {
+        if let Some(Ok(handle)) = self.handle.lock().get() {
             handle.state()
         } else {
             PlaybackState::Stopped
@@ -221,57 +353,62 @@ impl Sound {
     }
     /// Returns the playback position in seconds.
     pub fn position(&self) -> f64 {
-        if let Some(handle) = self.handle.lock().get() {
+        if let Some(Ok(handle)) = self.handle.lock().get() {
             handle.position()
         } else {
             0.0
         }
     }
     /// Sets the volume of the sound.
-    pub fn set_volume(&mut self, volume: impl Into<Volume>, tween: Tween) {
+    pub fn set_volume(&mut self, volume: impl Into<Volume>, tween: Tween) -> Result<()> {
         let volume = volume.into();
         let value_volume = Value::Fixed(volume);
         self.settings.volume = volume;
-        if let Some(handle) = self.handle.lock().get_mut() {
-            handle.set_volume(value_volume, tween).unwrap();
+        if let Some(Ok(handle)) = self.handle.lock().get_mut() {
+            handle.set_volume(value_volume, tween.into())?;
         }
+        Ok(())
     }
     /// Sets the rate, at which the sound is getting played.
-    pub fn set_playback_rate(&mut self, playback_rate: impl Into<PlaybackRate>, tween: Tween) {
+    pub fn set_playback_rate(
+        &mut self,
+        playback_rate: impl Into<PlaybackRate>,
+        tween: Tween,
+    ) -> Result<()> {
         let playback_rate = playback_rate.into();
         let value_playback_rate = Value::Fixed(playback_rate);
         self.settings.playback_rate = playback_rate;
-        if let Some(handle) = self.handle.lock().get_mut() {
-            handle
-                .set_playback_rate(value_playback_rate, tween)
-                .unwrap();
+        if let Some(Ok(handle)) = self.handle.lock().get_mut() {
+            handle.set_playback_rate(value_playback_rate, tween.into())?;
         }
+        Ok(())
     }
     /// Sets the panning of the sound, where 0.0 is left and 1.0 is right.
-    pub fn set_panning(&mut self, panning: f64, tween: Tween) {
+    pub fn set_panning(&mut self, panning: f64, tween: Tween) -> Result<()> {
         let value_panning = Value::Fixed(panning);
         self.settings.panning = panning;
-        if let Some(handle) = self.handle.lock().get_mut() {
-            handle.set_panning(value_panning, tween).unwrap();
+        if let Some(Ok(handle)) = self.handle.lock().get_mut() {
+            handle.set_panning(value_panning, tween.into())?;
         }
+        Ok(())
     }
     /// Sets the region, where the sound is getting played.
-    pub fn set_playback_region(&mut self, playback_region: impl Into<Region>) {
+    pub fn set_playback_region(&mut self, playback_region: impl Into<Region>) -> Result<()> {
         self.settings.playback_region = playback_region.into();
-        if let Some(handle) = self.handle.lock().get_mut() {
-            handle
-                .set_playback_region(self.settings.playback_region)
-                .unwrap();
+        if let Some(Ok(handle)) = self.handle.lock().get_mut() {
+            handle.set_playback_region(self.settings.playback_region)?;
         }
+        Ok(())
     }
     /// Sets the optional region, where the sound is getting looped.
-    pub fn set_loop_region(&mut self, loop_region: impl IntoOptionalRegion) {
+    pub fn set_loop_region(&mut self, loop_region: impl IntoOptionalRegion) -> Result<()> {
         self.settings.loop_region = loop_region.into_optional_loop_region();
-        if let Some(handle) = self.handle.lock().get_mut() {
-            handle.set_loop_region(self.settings.loop_region).unwrap();
+        if let Some(Ok(handle)) = self.handle.lock().get_mut() {
+            handle.set_loop_region(self.settings.loop_region)?;
         }
+        Ok(())
     }
-    /// Binds an object to this sound and plays it where this object is at.
+    /// Binds an object to this sound and plays it where this object is located at.
     pub fn bind_to_object(&mut self, object: Option<&Object>) {
         self.object = object.cloned();
     }
@@ -280,59 +417,63 @@ impl Sound {
         self.object.as_ref()
     }
     /// Updates the position of the sound.
-    pub fn update(&mut self) {
+    pub fn update(&mut self, tween: Tween) -> Result<()> {
         if let (Some(emitter), Some(object)) = (self.emitter.lock().get_mut(), &mut self.object) {
             object.update();
-            emitter
-                .set_position(object.transform.position.extend(0.0), Tween::default())
-                .unwrap()
+            emitter.set_position(object.transform.position.extend(0.0), tween.into())?
         }
+        Ok(())
     }
 
     /// Plays this sound.
-    pub fn play(&mut self) {
+    pub fn play(&mut self) -> Result<()> {
         if self.state() != PlaybackState::Playing {
             RESOURCES
                 .audio_server
-                .send(AudioUpdate::Play(self.clone()))
-                .unwrap();
+                .send(AudioUpdate::Play(self.clone()))?;
         }
+        Ok(())
     }
     /// Pauses this sound.
-    pub fn pause(&mut self, tween: Tween) {
+    pub fn pause(&mut self, tween: Tween) -> Result<()> {
         if self.state() != PlaybackState::Paused {
-            if let Some(handle) = self.handle.lock().get_mut() {
-                handle.pause(tween);
+            if let Some(Ok(handle)) = self.handle.lock().get_mut() {
+                handle.pause(tween.into())?;
             }
         }
+        Ok(())
     }
     /// Resumes the playback of this sound.
-    pub fn resume(&mut self, tween: Tween) {
+    pub fn resume(&mut self, tween: Tween) -> Result<()> {
         if self.state() != PlaybackState::Playing {
-            if let Some(handle) = self.handle.lock().get_mut() {
-                handle.resume(tween);
+            if let Some(Ok(handle)) = self.handle.lock().get_mut() {
+                handle.resume(tween.into())?;
             }
         }
+        Ok(())
     }
     /// Stops this sound.
-    pub fn stop(&mut self, tween: Tween) {
+    pub fn stop(&mut self, tween: Tween) -> Result<()> {
         if self.state() != PlaybackState::Stopped {
-            if let Some(handle) = self.handle.lock().get_mut() {
-                handle.stop(tween);
+            if let Some(Ok(handle)) = self.handle.lock().get_mut() {
+                handle.stop(tween.into())?;
             }
         }
+        Ok(())
     }
     /// Sets the playhead to the given position in seconds.
-    pub fn seek_to(&mut self, position: f64) {
-        if let Some(handle) = self.handle.lock().get_mut() {
-            handle.seek_to(position);
+    pub fn seek_to(&mut self, position: f64) -> Result<()> {
+        if let Some(Ok(handle)) = self.handle.lock().get_mut() {
+            handle.seek_to(position)?;
         }
+        Ok(())
     }
     /// Sets the playhead ahead by the given seconds.
-    pub fn seek_by(&mut self, position: f64) {
-        if let Some(handle) = self.handle.lock().get_mut() {
-            handle.seek_by(position);
+    pub fn seek_by(&mut self, position: f64) -> Result<()> {
+        if let Some(Ok(handle)) = self.handle.lock().get_mut() {
+            handle.seek_by(position)?;
         }
+        Ok(())
     }
 }
 
@@ -385,7 +526,7 @@ impl SoundSettings {
             volume,
             playback_rate,
             panning,
-            fade_in_tween: settings.fade_in_tween,
+            fade_in_tween: settings.fade_in_tween.map(Tween::from),
         }
     }
     builder_pattern!(playback_region, "the playback region", Region);
@@ -406,7 +547,7 @@ impl From<SoundSettings> for StaticSoundSettings {
             .volume(value.volume)
             .playback_rate(value.playback_rate)
             .panning(value.panning)
-            .fade_in_tween(value.fade_in_tween)
+            .fade_in_tween(value.fade_in_tween.map(kira::tween::Tween::from))
     }
 }
 
@@ -427,7 +568,7 @@ impl From<StaticSoundSettings> for SoundSettings {
             .volume(volume)
             .playback_rate(playback_rate)
             .panning(panning)
-            .fade_in_tween(value.fade_in_tween)
+            .fade_in_tween(value.fade_in_tween.map(Tween::from))
     }
 }